+++ toolchains/python-uv/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Ensures that Python is installed before setting up this toolchain.
+
 ## 0.1.4
 
 #### 🚀 Updates